requirements.txt
requests-oauthlib
pyyaml

